+++ Kernel/include/syscallDispatcher.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-uint64_t syscallSelector(uint64_t rdi, uint64_t rsi, uint64_t rdx, uint64_t rcx, uint64_t r8, uint64_t r9);
+int syscallSelector(uint64_t rdi, uint64_t rsi, uint64_t rdx, uint64_t rcx, uint64_t r8, uint64_t r9);
 
 #endif
